Fix N+1 selects problem for users
Even though FetchType.EAGER is set, the ORM doesn't seem to want to use a JOIN and instead loads permissions one-by-one. This should work around it.

A better solution would be to write a custom SQL query using a FETCH, but I can't be bothered to figure that out.

This should significantly improve loading times for the /admin and /configure/{circleId} endpoints